requirements.txt
django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
drf-yasg==1.21.7
django-cors-headers==4.4.0
django-filter==24.2
pillow==10.4.0
isort==5.13.2
